fix: change dmgValue type to float

+++ Assets/Scripts/DamageSystem/Damage.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public struct Damage
 {
-    public int DmgValue;
+    public float DmgValue;
 }
